+++ src/includes/wp-sharks-core-rv.php
@@ -12,6 +12,6 @@ if (!defined('WPINC')) {
     exit('Do NOT access this file directly.');
 }
 $GLOBALS['wp_sharks_core_rv'] = [
-    'min' => '170219.18660', //wp-sharks-core-required-version//
+    'min' => '170406.31578', //wp-sharks-core-required-version//
     'max' => '', //wp-sharks-core-max-compatible-version//
 ];